TST: Refactor test_parsers.py
Refactored tests in test_parsers.py
to increase coverage of the different
types of parsers and remove nearly
duplicate testing in some cases.
