+++ pandas/io/tests/parser/data/salary.table
@@ -1,0 +1,47 @@
+S	X	E	M
+13876 	1 	1 	1 
+11608 	1 	3 	0 
+18701 	1 	3 	1 
+11283 	1 	2 	0 
+11767 	1 	3 	0 
+20872 	2 	2 	1 
+11772 	2 	2 	0 
+10535 	2 	1 	0 
+12195 	2 	3 	0 
+12313 	3 	2 	0 
+14975 	3 	1 	1 
+21371 	3 	2 	1 
+19800 	3 	3 	1 
+11417 	4 	1 	0 
+20263 	4 	3 	1 
+13231 	4 	3 	0 
+12884 	4 	2 	0 
+13245 	5 	2 	0 
+13677 	5 	3 	0 
+15965 	5 	1 	1 
+12336 	6 	1 	0 
+21352 	6 	3 	1 
+13839 	6 	2 	0 
+22884 	6 	2 	1 
+16978 	7 	1 	1 
+14803 	8 	2 	0 
+17404 	8 	1 	1 
+22184 	8 	3 	1 
+13548 	8 	1 	0 
+14467 	10 	1 	0 
+15942 	10 	2 	0 
+23174 	10 	3 	1 
+23780 	10 	2 	1 
+25410 	11 	2 	1 
+14861 	11 	1 	0 
+16882 	12 	2 	0 
+24170 	12 	3 	1 
+15990 	13 	1 	0 
+26330 	13 	2 	1 
+17949 	14 	2 	0 
+25685 	15 	3 	1 
+27837 	16 	2 	1 
+18838 	16 	2 	0 
+17483 	16 	1 	0 
+19207 	17 	2 	0 
+19346 	20 	1 	0
