+++ pandas/io/tests/parser/data/test1.csv
@@ -1,0 +1,8 @@
+index,A,B,C,D
+2000-01-03 00:00:00,0.980268513777,3.68573087906,-0.364216805298,-1.15973806169
+2000-01-04 00:00:00,1.04791624281,-0.0412318367011,-0.16181208307,0.212549316967
+2000-01-05 00:00:00,0.498580885705,0.731167677815,-0.537677223318,1.34627041952
+2000-01-06 00:00:00,1.12020151869,1.56762092543,0.00364077397681,0.67525259227
+2000-01-07 00:00:00,-0.487094399463,0.571454623474,-1.6116394093,0.103468562917
+2000-01-10 00:00:00,0.836648671666,0.246461918642,0.588542635376,1.0627820613
+2000-01-11 00:00:00,-0.157160753327,1.34030689438,1.19577795622,-1.09700699751
