+++ pandas/io/tests/parser/data/test1.csv.gz
