+++ pandas/io/tests/parser/data/unicode_series.csv
