+++ pandas/tests/test_lib.py
@@ -1,12 +1,18 @@
 # -*- coding: utf-8 -*-
+
 from datetime import datetime, timedelta, date, time
 
 import numpy as np
-
 import pandas as pd
 import pandas.lib as lib
 import pandas.util.testing as tm
-from pandas.compat import u, PY2
+
+from pandas.compat import long, u, PY2
+
+
+def _assert_same_values_and_dtype(res, exp):
+    tm.assert_equal(res.dtype, exp.dtype)
+    tm.assert_almost_equal(res, exp)
 
 
 class TestMisc(tm.TestCase):
@@ -248,6 +254,71 @@ class Testisscalar(tm.TestCase):
         self.assertFalse(lib.isscalar(pd.Index([])))
         self.assertFalse(lib.isscalar(pd.Index([1])))
 
+
+class TestParseSQL(tm.TestCase):
+
+    def test_convert_sql_column_floats(self):
+        arr = np.array([1.5, None, 3, 4.2], dtype=object)
+        result = lib.convert_sql_column(arr)
+        expected = np.array([1.5, np.nan, 3, 4.2], dtype='f8')
+        _assert_same_values_and_dtype(result, expected)
+
+    def test_convert_sql_column_strings(self):
+        arr = np.array(['1.5', None, '3', '4.2'], dtype=object)
+        result = lib.convert_sql_column(arr)
+        expected = np.array(['1.5', np.nan, '3', '4.2'], dtype=object)
+        _assert_same_values_and_dtype(result, expected)
+
+    def test_convert_sql_column_unicode(self):
+        arr = np.array([u('1.5'), None, u('3'), u('4.2')],
+                       dtype=object)
+        result = lib.convert_sql_column(arr)
+        expected = np.array([u('1.5'), np.nan, u('3'), u('4.2')],
+                            dtype=object)
+        _assert_same_values_and_dtype(result, expected)
+
+    def test_convert_sql_column_ints(self):
+        arr = np.array([1, 2, 3, 4], dtype='O')
+        arr2 = np.array([1, 2, 3, 4], dtype='i4').astype('O')
+        result = lib.convert_sql_column(arr)
+        result2 = lib.convert_sql_column(arr2)
+        expected = np.array([1, 2, 3, 4], dtype='i8')
+        _assert_same_values_and_dtype(result, expected)
+        _assert_same_values_and_dtype(result2, expected)
+
+        arr = np.array([1, 2, 3, None, 4], dtype='O')
+        result = lib.convert_sql_column(arr)
+        expected = np.array([1, 2, 3, np.nan, 4], dtype='f8')
+        _assert_same_values_and_dtype(result, expected)
+
+    def test_convert_sql_column_longs(self):
+        arr = np.array([long(1), long(2), long(3), long(4)], dtype='O')
+        result = lib.convert_sql_column(arr)
+        expected = np.array([1, 2, 3, 4], dtype='i8')
+        _assert_same_values_and_dtype(result, expected)
+
+        arr = np.array([long(1), long(2), long(3), None, long(4)], dtype='O')
+        result = lib.convert_sql_column(arr)
+        expected = np.array([1, 2, 3, np.nan, 4], dtype='f8')
+        _assert_same_values_and_dtype(result, expected)
+
+    def test_convert_sql_column_bools(self):
+        arr = np.array([True, False, True, False], dtype='O')
+        result = lib.convert_sql_column(arr)
+        expected = np.array([True, False, True, False], dtype=bool)
+        _assert_same_values_and_dtype(result, expected)
+
+        arr = np.array([True, False, None, False], dtype='O')
+        result = lib.convert_sql_column(arr)
+        expected = np.array([True, False, np.nan, False], dtype=object)
+        _assert_same_values_and_dtype(result, expected)
+
+    def test_convert_sql_column_decimals(self):
+        from decimal import Decimal
+        arr = np.array([Decimal('1.5'), None, Decimal('3'), Decimal('4.2')])
+        result = lib.convert_sql_column(arr)
+        expected = np.array([1.5, np.nan, 3, 4.2], dtype='f8')
+        _assert_same_values_and_dtype(result, expected)
 
 if __name__ == '__main__':
     import nose
